requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-tinymce==3.4.0
gunicorn==20.1.0
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
